refactor: using latest node features
